+++ packages/autop/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Bug Fix
 
-- `autop` will correctly avoid adding `br` and `p` elements within `script` and `style` tags.
+- `autop` will correctly avoid adding `br` and `p` elements within `script`, `style`, and `svg` tags.
 
 ## 2.3.0 (2019-05-21)